Se implemento la funcionalidad de pagos y se corrigieron clases de deposito, retiro, pago y transferencia

+++ TECBank_BackEnd/Data_Input_Models/PagoDataInputModel.cs
@@ -8,7 +8,7 @@
 
         public string Apellido1 { get; set; }
         public string Apellido2 { get; set; }
-        public float Monto { get; set; }
+        public int Monto { get; set; }
         public string ID { get; set; }
         public string Moneda { get; set; }
         public String Cuenta_Emisora { get; set; }
